+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @manypkg/cli
 
+## 0.11.0
+
+### Minor Changes
+
+- [`d73628d`](https://github.com/Thinkmill/manypkg/commit/d73628db75a8ff3088cb4e62813dbf31b13b72bb) [#48](https://github.com/Thinkmill/manypkg/pull/48) Thanks [@mitchellhamilton](https://github.com/mitchellhamilton)! - Add repository field check which checks if a GitHub repo URL is in the `repository` field in the root `package.json` and if it is, checks that all of the packages have a `repository` field which goes into the directory of the package.
+
+## 0.10.1
+
+### Patch Changes
+
+- [`447c580`](https://github.com/Thinkmill/manypkg/commit/447c58010816c72a3a2a32504a284f9fb979dff1) [#45](https://github.com/Thinkmill/manypkg/pull/45) Thanks [@NateRadebaugh](https://github.com/NateRadebaugh)! - Add reference to `manypkg exec` to readme
+
+* [`c5275fb`](https://github.com/Thinkmill/manypkg/commit/c5275fba714c68539d304f07eb19962704fd560d) Thanks [@mitchellhamilton](https://github.com/mitchellhamilton)! - Show a more informative error when a package has no name
+
 ## 0.10.0
 
 ### Minor Changes